pyproject.toml
[project]
name = "off-policy"
description = "Off-policy RL (DDPG, TD3, SAC) algorithms from scratch"
version = "0.1.0"
authors = [{ name = "Akshay Raman" }]
license = "MIT"
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.7.1",
    "hydra-core>=1.3.2",
    "tqdm>=4.67.1",
    "numpy>=2.3.0",
    "gymnasium>=1.1.1",
]
